### 2024-04-30 #6707 @editorialbot
[PRE REVIEW]: exoTEDRF: An EXOplanet Transit and Eclipse Data Reduction Framework
**Submitting author:** @radicamc (<a href="http://orcid.org/0000-0002-3328-1203">Michael  Radica</a>)
**Repository:** https://github.com/radicamc/exoTEDRF
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.0
