### 2024-05-08 a533db9
Streamline whether files or data are passed
Philosophy is now that classes take files (or the data) and methods take data only.
Also add docstrings for class methods
