+++ exotedrf/stage2.py
@@ -34,17 +34,44 @@ class AssignWCSStep:
     """Wrapper around default calwebb_spec2 Assign WCS step.
     """
 
-    def __init__(self, datafiles, output_dir='./'):
+    def __init__(self, input_data, output_dir='./'):
         """Step initializer.
+
+        Parameters
+        ----------
+        input_data : array-like(str), array-like(datamodel)
+            List of paths to input data or the input data itself.
+        output_dir : str
+            Path to directory to which to save outputs.
         """
 
+        # Set up easy attributes.
         self.tag = 'assignwcsstep.fits'
         self.output_dir = output_dir
-        self.datafiles = np.atleast_1d(datafiles)
+
+        # Unpack input data files.
+        datafiles = utils.sort_datamodels(input_data)
+        self.datafiles = []
+        for file in datafiles:
+            self.datafiles.append(utils.open_filetype(file))
         self.fileroots = utils.get_filename_root(self.datafiles)
 
     def run(self, save_results=True, force_redo=False, **kwargs):
         """Method to run the step.
+
+        Parameters
+        ----------
+        save_results : bool
+            If True, save results.
+        force_redo : bool
+            If True, run step even if output files are detected.
+        kwargs : dict
+            Keyword arguments for calwebb_spec2.assign_wcs_step.AssignWcsStep.
+
+        Returns
+        -------
+        results : list(datamodel)
+            Input data files processed through the step.
         """
 
         results = []
@@ -55,7 +82,7 @@ class AssignWCSStep:
             if expected_file in all_files and force_redo is False:
                 fancyprint('File {} already exists.'.format(expected_file))
                 fancyprint('Skipping Assign WCS Step.')
-                res = expected_file
+                res = datamodels.open(expected_file)
             # If no output files are detected, run the step.
             else:
                 step = calwebb_spec2.assign_wcs_step.AssignWcsStep()
@@ -77,17 +104,44 @@ class SourceTypeStep:
     """Wrapper around default calwebb_spec2 Source Type Determination step.
     """
 
-    def __init__(self, datafiles, output_dir='./'):
+    def __init__(self, input_data, output_dir='./'):
         """Step initializer.
+
+        Parameters
+        ----------
+        input_data : array-like(str), array-like(datamodel)
+            List of paths to input data or the input data itself.
+        output_dir : str
+            Path to directory to which to save outputs.
         """
 
+        # Set up easy attributes.
         self.tag = 'sourcetypestep.fits'
         self.output_dir = output_dir
-        self.datafiles = np.atleast_1d(datafiles)
+
+        # Unpack input data files.
+        datafiles = utils.sort_datamodels(input_data)
+        self.datafiles = []
+        for file in datafiles:
+            self.datafiles.append(utils.open_filetype(file))
         self.fileroots = utils.get_filename_root(self.datafiles)
 
     def run(self, save_results=True, force_redo=False, **kwargs):
         """Method to run the step.
+
+        Parameters
+        ----------
+        save_results : bool
+            If True, save results.
+        force_redo : bool
+            If True, run step even if output files are detected.
+        kwargs : dict
+            Keyword arguments for calwebb_spec2.srctype_step.SourceTypeStep.
+
+        Returns
+        -------
+        results : list(datamodel)
+            Input data files processed through the step.
         """
 
         results = []
@@ -98,7 +152,7 @@ class SourceTypeStep:
             if expected_file in all_files and force_redo is False:
                 fancyprint('File {} already exists.'.format(expected_file))
                 fancyprint('Skipping Source Type Determination Step.')
-                res = expected_file
+                res = datamodels.open(expected_file)
             # If no output files are detected, run the step.
             else:
                 step = calwebb_spec2.srctype_step.SourceTypeStep()
[…]

[…]
