### 2024-05-09 fecb377
Add NIRSpec centroiding code
NOTE: Requires updated APPLESOSS v2.1.0
