+++ exotedrf/stage1.py
@@ -431,33 +431,34 @@ class OneOverFStep:
     """Wrapper around custom 1/f Correction Step.
     """
 
-    def __init__(self, input_data, baseline_ints, output_dir,
-                 method='scale-achromatic', timeseries=None,
-                 timeseries_o2=None, pixel_masks=None, background=None,
-                 centroids=None):
+    def __init__(self, input_data, output_dir, baseline_ints=None,
+                 pixel_masks=None, centroids=None, soss_background=None,
+                 method='scale-achromatic', soss_timeseries=None,
+                 soss_timeseries_o2=None):
         """Step initializer.
 
         Parameters
         ----------
         input_data : array-like(str), array-like(datamodel)
             List of paths to input data or the input data itself.
-        baseline_ints : array-like(int)
+        baseline_ints : array-like(int), None
             Integration number(s) to use as ingress and/or egress.
         output_dir : str
             Path to directory to which to save outputs.
         method : str
-            Correction method. Options are "scale-chromatic",
+            Correction method. SOSS options are "scale-chromatic",
             "scale-achromatic", "scale-achromatic-window", or "solve".
-        timeseries : np.ndarray(float), str, None
+            NIRSpec options are "median" or "slope".
+        soss_timeseries : np.ndarray(float), str, None
             Path to a file containing light curve(s) for order 1, or the
             light curve(s) themselves.
-        timeseries_o2 : np.ndarray(float), str, None
+        soss_timeseries_o2 : np.ndarray(float), str, None
             Path to a file containing light curves for order 2, or the
             light curves themselves.
         pixel_masks : array-like(str), np.ndarray(float), None
             List of paths to maps of pixels to mask for each data segment or
             the masks themselves. Should be 3D (nints, dimy, dimx).
-        background : np.ndarray(float), str, None
+        soss_background : np.ndarray(float), str, None
             Model of background flux.
         centroids : str, dict, None
             Path to file containing trace positions for each order or the
@@ -468,7 +469,6 @@ class OneOverFStep:
         self.tag = 'oneoverfstep.fits'
         self.output_dir = output_dir
         self.method = method
-        self.baseline_ints = baseline_ints
 
         # Unpack input data files.
         datafiles = utils.sort_datamodels(input_data)
@@ -485,25 +485,25 @@ class OneOverFStep:
             self.centroids = centroids
 
         # Unpack timeseries.
-        if isinstance(timeseries, str):
-            fancyprint('Reading timeseries file: {}...'.format(timeseries))
-            self.timeseries = np.load(timeseries)
-        elif isinstance(timeseries, np.ndarray) or timeseries is None:
-            self.timeseries = timeseries
+        if isinstance(soss_timeseries, str):
+            fancyprint('Reading timeseries file: {}...'.format(soss_timeseries))
+            self.timeseries = np.load(soss_timeseries)
+        elif isinstance(soss_timeseries, np.ndarray) or soss_timeseries is None:
+            self.timeseries = soss_timeseries
         else:
-            msg = 'Invalid type for timeseries: {}'.format(type(timeseries))
+            msg = 'Invalid type for timeseries: {}'.format(type(soss_timeseries))
             raise ValueError(msg)
 
         # Unpack timeseries for order 2.
-        if isinstance(timeseries_o2, str):
+        if isinstance(soss_timeseries_o2, str):
             fancyprint('Reading order 2timeseries file: {}...'
-                       ''.format(timeseries_o2))
-            self.timeseries_o2 = np.load(timeseries_o2)
-        elif isinstance(timeseries_o2, np.ndarray) or timeseries_o2 is None:
[…]
