### 2024-05-10 e4d43ed
NIRSpec stage 1 done.
