+++ exotedrf/stage2.py
@@ -100,6 +100,76 @@ class AssignWCSStep:
         return results
 
 
+class Extract2DStep:
+    """Wrapper around default calwebb_spec2 2D Extraction step.
+    """
+
+    def __init__(self, input_data, output_dir='./'):
+        """Step initializer.
+
+        Parameters
+        ----------
+        input_data : array-like(str), array-like(datamodel)
+            List of paths to input data or the input data itself.
+        output_dir : str
+            Path to directory to which to save outputs.
+        """
+
+        # Set up easy attributes.
+        self.tag = 'extract2dstep.fits'
+        self.output_dir = output_dir
+
+        # Unpack input data files.
+        datafiles = utils.sort_datamodels(input_data)
+        self.datafiles = []
+        for file in datafiles:
+            self.datafiles.append(utils.open_filetype(file))
+        self.fileroots = utils.get_filename_root(self.datafiles)
+
+    def run(self, save_results=True, force_redo=False, **kwargs):
+        """Method to run the step.
+
+        Parameters
+        ----------
+        save_results : bool
+            If True, save results.
+        force_redo : bool
+            If True, run step even if output files are detected.
+        kwargs : dict
+            Keyword arguments for calwebb_spec2.extract_2d_step.Extract2dStep.
+
+        Returns
+        -------
+        results : list(datamodel)
+            Input data files processed through the step.
+        """
+
+        results = []
+        all_files = glob.glob(self.output_dir + '*')
+        for i, segment in enumerate(self.datafiles):
+            # If an output file for this segment already exists, skip the step.
+            expected_file = self.output_dir + self.fileroots[i] + self.tag
+            if expected_file in all_files and force_redo is False:
+                fancyprint('File {} already exists.'.format(expected_file))
+                fancyprint('Skipping 2D Extraction Step.')
+                res = datamodels.open(expected_file)
+            # If no output files are detected, run the step.
+            else:
+                step = calwebb_spec2.extract_2d_step.Extract2dStep()
+                res = step.call(segment, output_dir=self.output_dir,
+                                save_results=save_results, **kwargs)
+                # Verify that filename is correct.
+                if save_results is True:
+                    current_name = self.output_dir + res.meta.filename
+                    if expected_file != current_name:
+                        res.close()
+                        os.rename(current_name, expected_file)
+                        res = datamodels.open(expected_file)
+            results.append(res)
+
+        return results
+
+
 class SourceTypeStep:
     """Wrapper around default calwebb_spec2 Source Type Determination step.
     """
@@ -156,6 +226,76 @@ class SourceTypeStep:
             # If no output files are detected, run the step.
             else:
                 step = calwebb_spec2.srctype_step.SourceTypeStep()
+                res = step.call(segment, output_dir=self.output_dir,
+                                save_results=save_results, **kwargs)
+                # Verify that filename is correct.
+                if save_results is True:
+                    current_name = self.output_dir + res.meta.filename
+                    if expected_file != current_name:
+                        res.close()
+                        os.rename(current_name, expected_file)
+                        res = datamodels.open(expected_file)
+            results.append(res)
+
+        return results
+
+
+class WaveCorrStep:
+    """Wrapper around default calwebb_spec2 Wavelength Correction step.
+    """
+
+    def __init__(self, input_data, output_dir='./'):
[…]
