+++ exotedrf/stage3.py
@@ -134,16 +134,24 @@ class Extract1DStep:
         self.pl_name = self.target_name + ' ' + planet_letter
         self.stellar_params = [st_teff, st_logg, st_met]
 
-    def run(self, soss_width=40, specprofile=None, centroids=None,
+        # Get instrument.
+        self.instrument = utils.get_instrument_name(self.datafiles[0])
+        if self.instrument == 'NIRSPEC' and extract_method == 'atoca':
+            fancyprint('ATOCA extraction selected but observation does not '
+                       'use NIRISS/SOSS. Switching to box extraction.',
+                       msg_type='WARNING')
+            self.extract_method = 'box'
+
+    def run(self, extract_width=40, soss_specprofile=None, centroids=None,
             save_results=True, force_redo=False, do_plot=False,
             show_plot=False, use_pastasoss=False, soss_estimate=None):
         """Method to run the step.
 
         Parameters
         ----------
-        soss_width : int
+        extract_width : int
             Full width of extraction aperture to use.
-        specprofile : str, None
+        soss_specprofile : str, None
             Path to specprofile file.
         centroids : str, None
             Path to file containing centroids for each order.
@@ -181,93 +189,20 @@ class Extract1DStep:
             spectra = expected_file
         # If no output file is detected, run the step.
         else:
-            # Option 1: ATOCA extraction.
+            # Option 1: ATOCA extraction - SOSS only.
             if self.extract_method == 'atoca':
-                if specprofile is None:
+                if soss_specprofile is None:
                     raise ValueError('specprofile reference file must be '
                                      'provided for ATOCA extraction.')
-                results = []
-                to_extract = {}
-                first_time = True
-                for i, file in enumerate(self.datafiles):
-                    to_extract['{}'.format(i)] = file
-                while len(to_extract) != 0:
-                    extracted = []
-                    for i in to_extract.keys():
-                        segment = to_extract[i]
-                        # Initialize extraction parameters for ATOCA.
-                        soss_modelname = self.fileroots[int(i)][:-1]
-                        # Perform the extraction.
-                        step = calwebb_spec2.extract_1d_step.Extract1dStep()
-                        try:
-                            res = step.call(segment,
-                                            output_dir=self.output_dir,
-                                            save_results=save_results,
-                                            soss_transform=[0, 0, 0],
-                                            subtract_background=False,
-                                            soss_bad_pix='model',
-                                            soss_width=soss_width,
-                                            soss_modelname=soss_modelname,
-                                            override_specprofile=specprofile,
-                                            soss_estimate=soss_estimate)
-                            # Verify that filename is correct.
-                            if save_results is True:
-                                current_name = self.output_dir + res.meta.filename
-                                if expected_file != current_name:
-                                    res.close()
-                                    os.rename(current_name, expected_file)
-                                    res = datamodels.open(expected_file)
-                            results.append(res)
-                            # Note that this segment was extracted correctly.
-                            extracted.append(i)
-                            # The first time that an extraction is successful,
-                            # create a soss_estimate if one does not already
-                            # exist.
-                            if first_time is True and soss_estimate is None:
-                                atoca_spectra = self.output_dir + self.fileroots[int(i)] + 'AtocaSpectra.fits'
-                                soss_estimate = get_soss_estimate(atoca_spectra,
-                                                                  output_dir=self.output_dir)
-                                first_time = False
-                        # When using ATOCA, sometimes a very specific error
-                        # pops up when an initial estimate of the stellar
-                        # spectrum cannot be obtained. This is needed to
-                        # establish the wavelength grid (which has a varying
[…]
