### 2024-05-10 178a492
Update DMS config file
