### 2024-05-10 2d77b19
fix bug in light curve sigma clipping
