### 2024-05-17 236dccc
NIRSpec tutorial
