+++ docs/content/citations.rst
@@ -3,6 +3,8 @@ Citations
 
 If you make use of this code in your work, please cite `Radica et al. (2023) <https://ui.adsabs.harvard.edu/abs/2023MNRAS.524..835R/abstract>`_ and `Feinstein et al. (2023) <https://ui.adsabs.harvard.edu/abs/2023Natur.614..670F/abstract>`_.
 
+For NIRSpec capabilities, please also cite `PLACEHOLDER <xxx>`_.
+
 Additional Citations
 --------------------
 If you use the ATOCA extraction algorithm for NIRISS/SOSS, please also cite `Radica et al. (2022) <https://ui.adsabs.harvard.edu/abs/2022PASP..134j4502R/abstract>`_
