### 2024-05-21 7ce7b9f
Update stage3.py
