### 2024-06-04 #6707 @radicamc
My direct contributions to the Alderson and Rustamkulov papers are minimal enough that I don’t think there would be any major conflicts with the main contributors to those papers (with the exception of Lili Alderson & Evie Ahrer as we are currently collaborating on multiple projects). Another potential reviewer, who I can’t recall whether or not they were part of many of the early JWST papers, is Taylor Bell.

—————————————————————————
Michael Radica
Étudiant au doctorat | Ph.D Candidate
Institut de recherche sur les exoplanètes | Institute for Research on Exoplanets (iREx)
Université de Montréal
https://radicamc.github.io<https://radicamc.github.io/>
http://www.exoplanetes.umontreal.ca/equipe/etudiants/michael-radica/
