### 2024-06-17 #6898 @editorialbot
[REVIEW]: exoTEDRF: An EXOplanet Transit and Eclipse Data Reduction Framework
**Submitting author:** @radicamc (<a href="http://orcid.org/0000-0002-3328-1203">Michael  Radica</a>)
**Repository:** https://github.com/radicamc/exoTEDRF
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
