### 2024-06-28 radicamc/exoTEDRF#7 @mm-murphy
Minor Comments on JOSS Software Paper
Hey Michael, great work so far on exoTEDRF! This issue relates to checking your Software paper as part of my review for JOSS (https://github.com/openjournals/joss-reviews/issues/6898).

Overall, the paper is well written and concisely states the purpose of exoTEDRF, how it is used, and where users can go to find the code and its documentation. exoTEDRF is slotting into a much needed niche -- reducing NIRISS SOSS observations from JWST. I do have one minor comment on your paper related to this for the State of the Field section of my checklist... As you mention in the text and in the 'Similar Tools' section, there are a number of other pipelines for JWST data reduction. Each of these either serves as a general use pipeline or is specially tailored to one instrument or another, just like exoTEDRF in its current form. This is an important point when choosing one pipeline over another, but would not be obvious to newcomers, so I would recommend adding a couple words of explanation on this matter. For example, it would be really useful to outline which pipelines deal specifically with X instrument, which ones are capable of handling all the instruments, and how exoTEDRF differs from any other codes doing NIRISS/SOSS (if there are any). I believe this would really highlight how exoTEDRF fills a hole in the current set of reduction codes, as you allude to in the Statement of Need, in addition to helping future users decide to use exoTEDRF over the other options.  This isn't a _need to do_ but I do believe it would really be a benefit.

Also, I noticed one typo on line 56: "particulalry" should be "particularly".

Otherwise, I find that the paper checks all the necessary boxes!
