### 2024-06-28 radicamc/exoTEDRF#7 @radicamc
Hi Matt,
Yes great suggestion! I will certainly add that to the manuscript.
I'm also in the process of prepping a new release of the pipeline with NIRSpec support and which removes the dependencies on Juliet for the light curve fitting in stage 4 in favour of another more hackable code that I've written. I hope this should be ready by the end of next week. At that point I'll push a new version of the manuscript with the changes you suggest, as well as a couple of edits reflecting the fact that NIRSpec is now supported, etc.
