### 2024-07-04 0f2f1a5
Make Upporf and Teddorf play nicely. Take 1
