+++ exotedrf/plotting.py
@@ -11,7 +11,6 @@ Plotting routines.
 from astropy.io import fits
 from astropy.timeseries import LombScargle
 import bottleneck as bn
-import corner
 import matplotlib.backends.backend_pdf
 from matplotlib.gridspec import GridSpec
 from matplotlib.patches import Ellipse
@@ -20,8 +19,6 @@ import numpy as np
 from scipy.ndimage import median_filter
 from tqdm import tqdm
 import warnings
-
-from jwst import datamodels
 
 from exotedrf import utils
 from exotedrf.utils import fancyprint
@@ -286,34 +283,6 @@ def make_compare_spectra_plot(spec1, spec2, title=None):
     return dev
 
 
-def make_corner_plot(fit_params, results, posterior_names=None, outpdf=None,
-                     truths=None):
-    """Make corner plot for lightcurve fitting.
-    """
-
-    first_time = True
-    for param in fit_params:
-        if first_time:
-            pos = results.posteriors['posterior_samples'][param]
-            first_time = False
-        else:
-            pos = np.vstack((pos, results.posteriors['posterior_samples'][param]))
-
-    figure = corner.corner(pos.T, labels=posterior_names, color='black',
-                           show_titles=True, title_fmt='.3f',
-                           label_kwargs=dict(fontsize=14), truths=truths,
-                           facecolor='white')
-    if outpdf is not None:
-        if isinstance(outpdf, matplotlib.backends.backend_pdf.PdfPages):
-            outpdf.savefig(figure)
-        else:
-            figure.savefig(outpdf)
-        figure.clear()
-        plt.close(figure)
-    else:
-        plt.show()
-
-
 def make_decontamination_plot(results, models, outfile=None, show_plot=True):
     """Nine-pixel plot for ATOCA decontamination.
     """
@@ -432,151 +401,6 @@ def make_jump_location_plot(results, outfile=None, show_plot=True):
         fancyprint('Plot saved to {}'.format(outfile))
     if show_plot is False:
         plt.close()
-    else:
-        plt.show()
-
-
-def make_lightcurve_plot(t, data, model, scatter, errors, nfit, outpdf=None,
-                         title=None, systematics=None, rasterized=False,
-                         nbin=10):
-    """Plot results of lightcurve fit.
-    """
-
-    def gaus(x, m, s):
-        return np.exp(-0.5*(x - m)**2/s**2)/np.sqrt(2*np.pi*s**2)
-
-    def chi2(o, m, e):
-        return np.nansum((o - m)**2/e**2)
-
-    if systematics is not None:
-        fig = plt.figure(figsize=(13, 9), facecolor='white',
-                         rasterized=rasterized)
-        gs = GridSpec(5, 1, height_ratios=[3, 3, 1, 0.3, 1])
-    else:
-        fig = plt.figure(figsize=(13, 7), facecolor='white',
-                         rasterized=rasterized)
-        gs = GridSpec(4, 1, height_ratios=[3, 1, 0.3, 1])
-
-    # Light curve with full systematics + astrophysical model.
-    ax1 = plt.subplot(gs[0])
-    assert len(data) == len(model)
-    nint = len(data)  # Total number of data points
-    # Full dataset
-    ax1.errorbar(t, data, yerr=scatter*1e-6, fmt='o', capsize=0,
[…]
