### 2024-07-05 f88d89d
exouprf compatibility updates
