+++ exotedrf/fit_lightcurves.py
@@ -5,7 +5,7 @@ Created on Wed Jul 27 14:35 2022
 
 @author: MCR
 
-Juliet light curve fitting script.
+Script to fit simple transit and eclipse light curves with exoUPRF.
 """
 
 from astropy.io import fits
@@ -82,7 +82,7 @@ f.write('\nRun at {}.'.format(time))
 f.close()
 
 # Formatted parameter names for plotting.
-formatted_names = {'per_p1': r'$P$', 't0_p1': r'$T_0$', 'rp_p1': r'$R_p/R_*$',
+formatted_names = {'per_p1': r'$P$', 't0_p1': r'$T_0$', 'rp_p1_inst': r'$R_p/R_*$',
                    'inc_p1': r'$i$', 'u1_inst': r'$u_1$', 'u2_inst': r'$u_2$',
                    'ecc_p1': r'$e$', 'w_p1': r'$\Omega$',
                    'a_p1': r'$a/R_*$', 'sigma_inst': r'$\sigma$',
@@ -97,23 +97,22 @@ formatted_names = {'per_p1': r'$P$', 't0_p1': r'$T_0$', 'rp_p1': r'$R_p/R_*$',
 
 # === Get Detrending Quantities ===
 # Get time axis
-if config['observing_model'] == 'NIRISS/SOSS':
+if config['observing_mode'] == 'NIRISS/SOSS':
     t = fits.getdata(config['infile'], 9)
 else:
     t = fits.getdata(config['infile'], 5)
 # Quantities against which to linearly detrend.
 if config['lm_file'] is not None:
     lm_data = pd.read_csv(config['lm_file'], comment='#')
-    lm_quantities = np.zeros((len(t), len(config['lm_parameters'])+1))
-    lm_quantities[:, 0] = np.ones_like(t)
+    lm_quantities = np.zeros((len(config['lm_parameters'])+1, len(t)))
+    lm_quantities[0] = np.ones_like(t)
     for i, key in enumerate(config['lm_parameters']):
-        i += 1
         lm_param = lm_data[key]
-        lm_quantities[:, i] = (lm_param - np.mean(lm_param)) / np.sqrt(np.var(lm_param))
+        lm_quantities[i+1] = (lm_param - np.mean(lm_param)) / np.sqrt(np.var(lm_param))
 # Eclipses must fit for a baseline, which is done via the linear detrending.
 # So add this term to the fits if not already included.
 if config['lm_file'] is None and config['lc_model_type'] == 'eclipse':
-    lm_quantities = np.zeros((len(t), 1))
+    lm_quantities = np.zeros((1, len(t)))
     lm_quantities[:, 0] = np.ones_like(t)
     config['params'].append('theta0_inst')
     config['dists'].append('uniform')
@@ -279,7 +278,7 @@ for order in config['orders']:
         # For transit only; update the LD prior for this bin if available.
         # For prior: set prior width to 0.2 around the model value - based on
         # findings of Patel & Espinoza 2022.
-        if config['lc_model_type'] == 'transit':
+        if config['lc_model_type'] == 'transit' and config['ld_fit_type'] != 'free':
             if config['ld_model_type'] == 'quadratic-kipping':
                 low_lim = 0.0
             else:
@@ -350,20 +349,20 @@ for order in config['orders']:
         # If not skipped, append median and 1-sigma bounds.
         else:
             results_dict = fit_results[wavebin].get_results_from_fit()
-            if config['occultation_type'] == 'transit':
-                md = results_dict['rp_p1']['median']
-                up = results_dict['rp_p1']['up_1sigma']
-                lw = results_dict['rp_p1']['low_1sigma']
+            if config['lc_model_type'] == 'transit':
[…]
