+++ setup.py
@@ -18,8 +18,7 @@ setup(name='exotedrf',
                         'more_itertools', 'numpy', 'pandas', 'ray',
                         'scikit-learn', 'scipy', 'tqdm', 'pastasoss',
                         'pyyaml'],
-      extras_require={'stage4': ['exotedrf', 'dynesty==1.2.2', 'exotic_ld',
-                                 'h5py', 'juliet'],
+      extras_require={'stage4': ['exotedrf', 'exouprf', 'exotic_ld', 'h5py'],
                       'webbpsf': ['exotedrf', 'webbpsf>=1.1.1']},
       classifiers=[
         'Development Status :: 3 - Alpha',
