+++ exotedrf/fit_lightcurves.yaml
@@ -46,7 +46,7 @@ dists : ['fixed', 'fixed', 'uniform', 'fixed',
          'loguniform']
 values : [3.42525650, 2459751.821681146, [0.01, 0.9], 0.748,
           [0., 1.], [0., 1.], 0.0, 90., 8.82,
-          [0.1, 10000]]
+          [0.000001, 0.1]]
 
 # === Detrending Paramaters ===
 # Path to file containing linear detrending parameters.
