### 2024-07-05 a38c210
Fix bug in prior/fixed LD fitting
