### 2024-07-05 cc4ef72
Update fit_lightcurves.py
