+++ setup.py
@@ -4,7 +4,7 @@
 from setuptools import setup
 
 setup(name='exotedrf',
-      version='1.4.0',
+      version='2.0.0',
       license='MIT',
       author='Michael Radica',
       author_email='michael.radica@umontreal.ca',
