### 2024-07-05 cef8dfb
Update JOSS paper
