+++ joss/paper.md
@@ -13,7 +13,7 @@ authors:
 affiliations:
  - name: Trottier Institute for Research on Exoplanets (iREx), Université de Montréal, Montréal, Canada
    index: 1
-date: 30 April 2024
+date: 05 July 2024
 bibliography: paper.bib
 ---
 
@@ -25,7 +25,7 @@ it is imperative to ensure that the spectral features that drive atmosphere infe
 
 Here, I present the community with `exoTEDRF` (EXOplanet Transit and Eclipse Data Reduction Framework; formerly known as 
 `supreme-SPOON`), an end-to-end pipeline for data reduction and light curve analysis of time series observations (TSOs) of transiting exoplanets with JWST. The pipeline is highly modular and designed to produce reliable spectra from raw JWST exposures. 
-`exoTEDRF` consists of four stages, each of which are further subdivided into a series of steps. These steps can either be run individually, for example in a Jupyter notebook, or via the command line using the provided configuration files.
+`exoTEDRF` (pronounced exo-tedorf) consists of four stages, each of which are further subdivided into a series of steps. These steps can either be run individually, for example in a Jupyter notebook, or via the command line using the provided configuration files.
 The steps are highly tunable, allowing full control over every parameter in the reduction. Each step also produces diagnostic plots to allow the user to verify their results at each intermediate stage, and compare outputs with other pipelines if so desired.
 Finally, `exoTEDRF` has also been designed to be run in "batch" mode: simultaneously running multiple reductions, each tweaking a subset of parameters, to understand any impacts on the resulting atmosphere spectrum.
 
@@ -39,7 +39,7 @@ Like similar pipelines (`Eureka!` [@bell_eureka_2022], `jwst` [@bushouse_howard_
 - Stage 3, Spectral extraction: Extract the 2D stellar spectra from the 3D slope images.
 - Stage 4, Light curve fitting: An optional stage for the fitting of extracted light curves.
 
-In `exoTEDRF`, Stage 4 is an optional installation which is currently built around the excellent `juliet` library [@espinoza_juliet_2019], and incorporates tools such as `ExoTiC-LD` [@david_grant_2022_7437681] for the estimation of stellar limb darkening parameters. 
+In `exoTEDRF`, Stage 4 is an optional installation which is currently built around the `exoUPRF` library [@michael_radica_2024_12628066], and incorporates tools such as `ExoTiC-LD` [@david_grant_2022_7437681] for the estimation of stellar limb darkening parameters. 
 In certain places (e.g., superbias subtraction, flat field correction), `exoTEDRF` simply provides a wrapper around the existing functionalities of the `jwst` package maintained by the Space Telescope Science Institute. 
 
 
@@ -47,7 +47,7 @@ In certain places (e.g., superbias subtraction, flat field correction), `exoTEDR
 Data analysis is a challenging process that is encountered by all observational studies. Ensuring that the resulting 
 atmosphere spectra are robust against particular choices made in the reduction process is critical, especially as we push to characterize the atmospheres of small rocky planets. 
 The modularity and tunability of `exoTEDRF` make it easy to run multiple reductions of a given dataset, and therefore robustly ascertain whether the spectral features driving atmosphere inferences are robust, or sensitive to the peculiarities of a given reduction.
[…]
