+++ joss/paper.bib
@@ -398,3 +398,31 @@ archivePrefix = {arXiv},
     adsurl  = {https://rdcu.be/b08Wh},
     doi     = {10.1038/s41592-019-0686-2},
 }
+
+@ARTICLE{Cadieux2024,
+       author = {{Cadieux}, Charles and {Doyon}, Ren{\'e} and {MacDonald}, Ryan J. and {Turbet}, Martin and {Artigau}, {\'E}tienne and {Lim}, Olivia and {Radica}, Michael and {Fauchez}, Thomas J. and {Salhi}, Salma and {Dang}, Lisa and {Albert}, Lo{\"\i}c and {Coulombe}, Louis-Philippe and {Cowan}, Nicolas B. and {Lafreni{\`e}re}, David and {L'Heureux}, Alexandrine and {Piaulet}, Caroline and {Benneke}, Bj{\"o}rn and {Cloutier}, Ryan and {Charnay}, Benjamin and {Cook}, Neil J. and {Fournier-Tondreau}, Marylou and {Plotnykov}, Mykhaylo and {Valencia}, Diana},
+        title = "{Transmission Spectroscopy of the Habitable Zone Exoplanet LHS 1140 b with JWST/NIRISS}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Earth and Planetary Astrophysics},
+         year = 2024,
+        month = jun,
+          eid = {arXiv:2406.15136},
+        pages = {arXiv:2406.15136},
+          doi = {10.48550/arXiv.2406.15136},
+archivePrefix = {arXiv},
+       eprint = {2406.15136},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv240615136C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@software{michael_radica_2024_12628066,
+  author       = {Michael Radica},
+  title        = {radicamc/exoUPRF: exoUPRF v1.0.0},
+  month        = jul,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {v1.0.0},
+  doi          = {10.5281/zenodo.12628066},
+  url          = {https://doi.org/10.5281/zenodo.12628066}
+}
