### 2024-07-05 radicamc/exoTEDRF#7 @radicamc
A new version of the JOSS paper should now be available reflecting the changes I've made to the code and incorporating your suggestions!
