### 2024-07-05 #6898 @radicamc
Hi all!
I've made some updates to the submission reflecting @mm-murphy's suggestions as well as some updates I've made to the code (e.g., NIRSpec is now fully supported instead of in development, etc.).
