### 2024-07-11 radicamc/exoTEDRF#9 @radicamc
Ah yes, I believe I know what's going on here.
Can you confirm that when you say you're using "rate.fits" files that you're using the Stage 1 products downloaded from MAST, or something equivalent? If so, then yes, this is what's causing the problem. The rate.fits files are averages over all integrations (so its a 2D -- x pixel, y pixel -- file). The step is expecting a TSO (so 3D -- integrations, x pixel, y pixel -- file). The "rateints.fits" files are the ones you would want to use in this case then (again assuming you're getting the Stage 1 products from MAST).
