### 2024-07-13 radicamc/exoTEDRF#9 @mm-murphy
Ah, my mistake! Yes, you're right, and thank you for explaining. When I instead start from the rateints files everything runs smoothly.
