### 2024-07-19 radicamc/exoTEDRF#10 @mm-murphy
Typo in NIRISS/SOSS Stage 3 Tutorial Notebook
Note that this issue is connected to the ongoing JOSS review (https://github.com/openjournals/joss-reviews/issues/6898)

I believe you have a typo or accidentally forgot to update some old code in the Stage 3 - 1D Spectral Extraction section of the NIRISS/SOSS Reduction tutorial notebook. In the tutorial, the line where stage 3's Extract1DStep is ran says `results = step.run(soss_width=30, centroids=centroids, save_results=True, force_redo=True)`. Running this throws a TypeError which states that the first argument `soss_width` was unexpected. Looking at the API documentation for this step, I believe this argument is supposed to say `extract_width`.  After switching the argument name to `extract_width`, the issue is resolved, and the stage completes as expected.
