### 2024-07-25 radicamc/exoTEDRF#11 @karllark
Possible mismatch between jwst==1.12.5 and current crds reference files
I am working on my JOSS review. [ joss-reviews/issues#6898](https://github.com/openjournals/joss-reviews/issues/6898)

I am running the tutorial_niriss-soss notebook and have gotten an error in the superbias step.  Given the this package requires an older jwst pipeline version (1.12.5) I am wondering if this is a mismatch with the current crds reference files.  The error is below.  Or maybe this is a different issue - please advise.  I tried to use jwst==1.15.1, but get the same error.

```
2024-07-25 13:31:23,358 - stpipe.SuperBiasStep - INFO - SuperBiasStep instance created.
2024-07-25 13:31:23,367 - stpipe.SuperBiasStep - INFO - SuperBiasStep instance created.
2024-07-25 13:31:23,445 - stpipe.SuperBiasStep - INFO - Step SuperBiasStep running with args (<RampModel(158, 9, 256, 2048) from jw01366001001_04101_00001-seg001_nis_saturationstep.fits>,).
2024-07-25 13:31:23,446 - stpipe.SuperBiasStep - INFO - Step SuperBiasStep parameters are:
  pre_hooks: []
  post_hooks: []
  output_file: None
  output_dir: pipeline_outputs_directory/Stage1/
  output_ext: .fits
  output_use_model: False
  output_use_index: True
  save_results: True
  skip: False
  suffix: None
  search_output_file: True
  input_dir: ''
2024-07-25 13:31:23,463 - CRDS - INFO -  Fetching  ./crds_cache/references/jwst/niriss/jwst_niriss_superbias_0200.fits    6.3 M bytes  (1 / 1 files) (0 / 6.3 M bytes)
2024-07-25 13:31:25,155 - stpipe.SuperBiasStep - INFO - Using SUPERBIAS reference file ./crds_cache/references/jwst/niriss/jwst_niriss_superbias_0200.fits
UFuncTypeError                            Traceback (most recent call last)
Cell In[10], line 6
      1 step = stage1.SuperBiasStep(results, output_dir=outdir_s1)
      3 # This step has options for diagnostic plotting. Specifying do_plot=True will create the step
      4 # plot and save it to the output directory.
----> 6 results = step.run(save_results=True, force_redo=True, do_plot=True)

File ~/Python/other/exoTEDRF/exotedrf/stage1.py:263, in SuperBiasStep.run(self, save_results, force_redo, do_plot, show_plot, **kwargs)
    260 # If no output files are detected, run the step.
    261 else:
    262     step = calwebb_detector1.superbias_step.SuperBiasStep()
--> 263     res = step.call(segment, output_dir=self.output_dir,
    264                     save_results=save_results, **kwargs)
    265     # Verify that filename is correct.
    266     if save_results is True:

File ~/Bin/miniconda3/envs/jwst1.15.1/lib/python3.11/site-packages/stpipe/step.py:686, in Step.call(cls, *args, **kwargs)
    683 name = config.get("name", None)
    684 instance = cls.from_config_section(config, name=name, config_file=config_file)
--> 686 return instance.run(*args)

File ~/Bin/miniconda3/envs/jwst1.15.1/lib/python3.11/site-packages/stpipe/step.py:509, in Step.run(self, *args)
    507     self.prefetch(*args)
    508 try:
--> 509     step_result = self.process(*args)
    510 except TypeError as e:
    511     if "process() takes exactly" in str(e):

File ~/Bin/miniconda3/envs/jwst1.15.1/lib/python3.11/site-packages/jwst/superbias/superbias_step.py:44, in SuperBiasStep.process(self, input)
     41 bias_model = datamodels.SuperBiasModel(self.bias_name)
     43 # Do the bias subtraction
---> 44 result = bias_sub.do_correction(input_model, bias_model)
     46 # Close the superbias reference file model and
     47 # set the step status to complete
     48 bias_model.close()

File ~/Bin/miniconda3/envs/jwst1.15.1/lib/python3.11/site-packages/jwst/superbias/bias_sub.py:43, in do_correction(input_model, bias_model)
     40 bias_model.data[np.isnan(bias_model.data)] = 0.0
     42 # Subtract the bias ref image from the science data
---> 43 output_model = subtract_bias(input_model, bias_model)
     45 output_model.meta.cal_step.superbias = 'COMPLETE'
     47 return output_model

File ~/Bin/miniconda3/envs/jwst1.15.1/lib/python3.11/site-packages/jwst/superbias/bias_sub.py:80, in subtract_bias(input, bias)
     76 output.pixeldq = np.bitwise_or(input.pixeldq, bias.dq)
     78 # Subtract the superbias image from all groups and integrations
     79 # of the science data
---> 80 output.data -= bias.data
     82 # If ZEROFRAME is present, subtract the super bias.  Zero values
     83 # indicate bad data, so should be kept zero.
     84 if input.meta.exposure.zero_frame:

UFuncTypeError: Cannot cast ufunc 'subtract' output from dtype('float32') to dtype('uint16') with casting rule 'same_kind'
```
