### 2024-07-25 radicamc/exoTEDRF#11 @radicamc
Yes I've encountered this a bunch in the last month. I'm not exactly sure why it's happening -- my assumption is that something was changed on the STScI side which had some unintended consequences. I'm digging into it.
In any case, there's a very simple workaround. You just need to create a list with the paths to the output files from the SaturationStep, so e.g.,
`results = ['XXXXXX_seg001_saturationstep.fits' , 'XXXXX_seg002_saturationstep.fits, ...]`
and pass this as the input to the SuperbiasStep. Then everything will work fine.
