### 2024-07-25 radicamc/exoTEDRF#11 @karllark
Ok.  I will try this.

What is your plan for fixing this issue?  Putting in the code above or digging and finding out what changed?
