### 2024-07-25 radicamc/exoTEDRF#11 @radicamc
A combination of both. I'll push a note into the docs with the temporary workaround while I continue looking into it. Though, as I said previously, there likely isn't much I can do if its a crds reference file compatibility issue. We'd have to wait until its patched on the STScI end.
