### 2024-07-25 radicamc/exoTEDRF#11 @radicamc
Added note to tutorials in [eb3e8b9](https://github.com/radicamc/exoTEDRF/commit/eb3e8b960ed3a64e91f11494144055a4561f40c0).
