+++ README.md
@@ -14,7 +14,7 @@ Currently supports: **NIRISS/SOSS** and **NIRSpec/BOTS**. **MIRI/LRS** is in dev
 An installation guide, code documentation, and tutorials are included in the [exoTEDRF documentation](https://exotedrf.readthedocs.io/en/latest/index.html).
 
 ## Citations
-If you make use of this code in your work, please cite [Radica (2024)](), as well as [Radica et al. (2023)](https://ui.adsabs.harvard.edu/abs/2023MNRAS.524..835R/abstract) 
+If you make use of this code in your work, please cite [Radica (2024)](https://ui.adsabs.harvard.edu/abs/2024arXiv240717541R/abstract), as well as [Radica et al. (2023)](https://ui.adsabs.harvard.edu/abs/2023MNRAS.524..835R/abstract) 
 and [Feinstein et al. (2023)](https://ui.adsabs.harvard.edu/abs/2023Natur.614..670F/abstract) for its first uses.  
 
 A list of other works you should consider citing depending on which functionalities you are using can be found [here](https://exotedrf.readthedocs.io/en/latest/content/citations.html).
