### 2024-07-29 radicamc/exoTEDRF#12 @karllark
Community contributions?
As part of reviewing the joss paper (https://github.com/openjournals/joss-reviews/issues/6898) I did not see any guidelines for community contributions in the repository (or paper).  Are such possible and/or encouraged?  It would be good to have a clear statement of this in the README at a minimum.
