### 2024-07-29 radicamc/exoTEDRF#13 @karllark
Statement of Need in Documentation
As part of reviewing the joss paper (https://github.com/openjournals/joss-reviews/issues/6898). there is no statement of need in the documentation.  There is one in the paper and I would recommend using that in the docs themselves to provide users with this information.  Would be good to state clearly in the docs why the jwst pipeline is not sufficient (for example).
