### 2024-07-29 radicamc/exoTEDRF#14 @karllark
Memory requirements?
In performing my JOSS review (https://github.com/openjournals/joss-reviews/issues/6898), I have having issues running the NIRISS notebook.  I keep running out of memory on my 64 Gb machine.  Is this expected?  What are the memory requirements for this notebook/reducing TSO data.  I'm not an exoplanet researcher so maybe this is obvious if I was, regardless might be good to document this at least.
