### 2024-07-30 radicamc/exoTEDRF#14 @radicamc
Just to clarify, you mean memory as in RAM, and not storage space correct?
In any case, TSO data, particularly from NIRISS is large. Individual segments are often >8GB. So if its storage that you mean, then I'm not surprised that you run out after completing the notebook.
I've never run into memory issues, though I normally (and I think this is standard practice) reduce my observations on a cluster as opposed to a laptop. What step is failing for you, and I can look into it?
