### 2024-07-30 radicamc/exoTEDRF#14 @karllark
Yes, I mean RAM.  A number of the steps are causing the kernel to crash.  I'll try the NIRSpec notebook next.  All looks good, but would be fun to get one of the notebooks to finish.  Maybe you can come up with a smaller example, that would likely be good for other users as well as myself and my "small" amount of RAM.  Hah!
