### 2024-07-30 radicamc/exoTEDRF#11 @karllark
The note is great.  You might even consider having an example with exactly what is needed so it is easy to copy and paste.
