### 2024-08-02 613a884
Add info for contributions to docs/README
