+++ docs/content/contributions.rst
@@ -1,0 +1,14 @@
+Citations
+==========
+
+Whether you just want a bug fixed or want to design a whole new feature, here you'll find some useful information if you'd like to contribute to **exoTEDRF**.
+
+If you find a bug in the code, or have a request for a feature which is not yet implemented, please feel free to raise an issue on `GitHub <https://github.com/radicamc/exoTEDRF/issues>`_.
+
+When you install **exoTEDRF**, whether through GitHub or pip, you'll have all the content on the :code:`main` branch -- the current official code release.
+Like many open source projects, all new feature developments, bug fixes, etc. happen on the :code:`dev` branch, and are rigorously tested before being merged into the :code:`main` code.
+If you'd like to create a new feature in **exoTEDRF**, you can either fork the repository, or create a new branch off of :code:`dev`, where you can work on your improvements without worrying about breaking anything in the main code release!
+
+The one main exception to this workflow is for major bug fixes, which will be implemented directly in :code:`main`. However, I hope these will be few and far between!
+
+Thanks in advance for your contributions, regardless of how small they might be!
