### 2024-08-02 radicamc/exoTEDRF#12 @radicamc
Added some text to this effect in the README and docs in [613a884](https://github.com/radicamc/exoTEDRF/commit/613a884129e93c6c36067f10a57a412dfa5f4b04).
