### 2024-08-02 radicamc/exoTEDRF#13 @radicamc
Addressed in [f1e9c70](https://github.com/radicamc/exoTEDRF/commit/f1e9c7095c73e3d6cb31a20aac98a6b3b0fd6c35).
