### 2024-08-02 radicamc/exoTEDRF#14 @radicamc
Yeah, I have the same issue running on my 16GB laptop, some steps will require too much memory and crash.
Honestly, I think this is just an unavoidable facet of reducing TSO data, particularly from NIRISS which are the largest of all.
I'll make a note in the documentation that the memory requirements of a laptop are likely not going to be enough to run the notebook (and analyze any other observations).
I could, in theory, make a smaller dataset that could be reduced on a laptop. However, I don't think this would be worthwhile for a couple of reasons:
First, I've tried to structure the notebooks for this package (and my related exoUPRF package) such that a user is able to reproduce the WASP-39b results that we published last year. These are the first exoplanet results from JWST, and are extremely well known in the field. As such, I believe it has a lot of pedagogical value to use this data for the tutorial.
Second, as I mentioned above, in order to reduce any real observations, a user will most likely need to switch to a cluster. So I don't see too much value in testing out the code on your laptop with a fake dataset, then having to move to a cluster for real observations. In theory you'd need to re-run through the notebooks again to make sure everything works in the new installation.
I'll make a note to try and reduce the memory requirements wherever possible, such that, perhaps, things can be run on a laptop. But I don't think this is super high-priority.
Thoughts?
