### 2024-08-02 radicamc/exoTEDRF#14 @karllark
All good thoughts.  I do agree that reducing the memory use would be good.  I did notice that there was a case where the pipeline step ran fine, but then creating the diagnostic image/plot seemed to crash things.  Maybe this is just too much on top of already running the pipeline.  Or maybe this is keeping the large data in memory instead of using the disk.  Regardless, I agree with your points.

Mentioning the memory needs is a good and reasonable solution.
