### 2024-08-07 1091852
Fix small bug when fitting both SOSS orders in one run
Order 1 was not being written to the spectrum file.
