### 2024-08-07 #6898 @mm-murphy
After seeing the above addressed and closed issues, I've completed my review checklist @warrickball . exoTEDRF checks all the boxes; I have verified to the best of my ability that exoTEDRF functions as described, delivers solid results, and has helpful yet simple documentation. Great job, Michael! This reduction software will be invaluable to the community, and I look forward to being able to use it myself.
