### 2024-08-21 #6898 @karllark
With the last of bugs in the sample NIRISS notebook fixed, I was able to run the notebook from beginning to end.  This completes my review checklist @warrickball.   The exoTEDRF works as described and has sufficient documentation to pass my review.  Great to see the capabilities to modify the jwst pipeline be used so well.   Great job Michael!
