### 2024-08-23 #6898 @warrickball
Thanks @karllark and @mm-murphy for your reviews! @radicamc, I'll start working towards publishing the article now, starting with some tweaks to the paper I've made in PR https://github.com/radicamc/exoTEDRF/pull/17. We'll then create some archives to associate with the paper, which will complete the process.
