### 2024-08-24 be18903
Update Cadieux et al. (2024) to ApJL published version and fix one typo
