+++ joss/paper.bib
@@ -416,17 +416,19 @@ archivePrefix = {arXiv},
 @ARTICLE{Cadieux2024,
        author = {{Cadieux}, Charles and {Doyon}, Ren{\'e} and {MacDonald}, Ryan J. and {Turbet}, Martin and {Artigau}, {\'E}tienne and {Lim}, Olivia and {Radica}, Michael and {Fauchez}, Thomas J. and {Salhi}, Salma and {Dang}, Lisa and {Albert}, Lo{\"\i}c and {Coulombe}, Louis-Philippe and {Cowan}, Nicolas B. and {Lafreni{\`e}re}, David and {L'Heureux}, Alexandrine and {Piaulet}, Caroline and {Benneke}, Bj{\"o}rn and {Cloutier}, Ryan and {Charnay}, Benjamin and {Cook}, Neil J. and {Fournier-Tondreau}, Marylou and {Plotnykov}, Mykhaylo and {Valencia}, Diana},
         title = "{Transmission Spectroscopy of the Habitable Zone Exoplanet LHS 1140 b with JWST/NIRISS}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Earth and Planetary Astrophysics},
+      journal = {The Astrophysical Journal Letters},
+     keywords = {Exoplanets, Habitable planets, Planetary atmospheres, Super Earths, Ocean planets, Mini Neptunes, M dwarf stars, Transmission spectroscopy, 498, 695, 1244, 1655, 1151, 1063, 982, 2133, Astrophysics - Earth and Planetary Astrophysics},
          year = 2024,
-        month = jun,
-          eid = {arXiv:2406.15136},
-        pages = {arXiv:2406.15136},
-          doi = {10.48550/arXiv.2406.15136},
+        month = jul,
+       volume = {970},
+       number = {1},
+          eid = {L2},
+        pages = {L2},
+          doi = {10.3847/2041-8213/ad5afa},
 archivePrefix = {arXiv},
        eprint = {2406.15136},
  primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv240615136C},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024ApJ...970L...2C},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
