### 2024-08-24 #6898 @warrickball
Thanks for the quick response @radicamc. The last outstanding thing is to upload an archive of the code repo to an archiving service like Zenodo, FigShare, etc. and provide the link here.

These archives are often made using [GitHub integrations](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) but you can also just download an archive of the code manually and upload that to the relevant service.
