### 2024-08-25 #6898 @warrickball
I presume it's [this Zenodo archive](https://zenodo.org/records/13370466)? If so, I'll attach the metadata (DOI, version, etc) here.

The GitHub integration has picked me up as an author via my PRs to tweak the paper. You're welcome (and I encourage you) to remove my name manually from the generated author list on Zenodo. You should be able to change this without having to generate a new archive.
