### 2024-08-25 #6898 @warrickball
I also see the Zenodo archive has somehow identified the licence as CC-BY 4.0, rather than the BSD 3-clause that is actually used in the repo.
