### 2024-08-26 #6898 @radicamc
Ah, thanks for catching that! I've updated the Zenodo author list and license.
