### 2024-08-27 #6898 @warrickball
Nope, but thanks for having a look! It's now up to the editor-in-chief or associate editor-in-chief to make final checks and publish the paper.

Thanks too to @karllark and @mm-murphy for their reviews. JOSS would be impossible without our reviewers!
