### 2024-08-30 #6898 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
